+++ CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [v0.10.1]
 ### Fixed
 - The maximum read length reported in the sub-title of `components.fastcat.read_length_plot()` being one bp too large.
 ### Changed
+- Data types in demo changed to dropdown list.
 - MSA now accepts start and end co-ordinates for partial sequence visualisation.
 
 ## [v0.10.0]